refactor: address PR review feedback for handle availability
- Remove fragile string parsing (includes()) from availability detection;
  use stable status field directly as the API contract
- Remove redundant isTaken field (always derivable from !isAvailable)
- Render API-provided messages in UI instead of hardcoded text
- Remove disruptive Alert.alert for loading state; use inline indicator
- Rename hook file to useCheckUserHandleAvailability.ts to match export
- Move availability feedback below the input field for better UX
- Ensure files end with newline
