+++ frontend/src/screens/auth/SignUpScreenFirst.tsx
@@ -18,7 +18,7 @@ import {
   launchImageLibrary,
   ImagePickerResponse,
 } from 'react-native-image-picker';
-import {useCheckUserHandleAvailability} from '@/src/hooks/useCheckUserHandle';
+import {useCheckUserHandleAvailability} from '@/src/hooks/useCheckUserHandleAvailability';
 import {useVerificationMailMutation} from '@/src/hooks/useMailVerification';
 import {useRegdMutation} from '@/src/hooks/useUserRegistration';
 // eslint-disable-next-line @typescript-eslint/no-require-imports
@@ -42,7 +42,7 @@ const SignupPageFirst = ({navigation}: SignUpScreenFirstProp) => {
 
   const userHandle = username.trim();
 
-  const {data: checkhandle, isLoading} =
+  const {data: handleAvailability, isLoading: isCheckingHandle} =
     useCheckUserHandleAvailability(userHandle);
   const {mutate: verifyEmailMutation, isPending: verifyEmailPending} =
     useVerificationMailMutation();
@@ -160,11 +160,8 @@ const SignupPageFirst = ({navigation}: SignUpScreenFirstProp) => {
     } else if (password.length < 6) {
       Alert.alert('Password must be at least of 6 length');
       return;
-    } else if (isLoading) {
-      Alert.alert('Please wait while we check the user handle');
-      return;
-    } else if (checkhandle?.isTaken) {
-      Alert.alert('User handle is already in use');
+    } else if (handleAvailability && !handleAvailability.isAvailable) {
+      Alert.alert('User handle is not available', 'Please choose a different handle.');
       return;
     }
 
@@ -387,22 +384,6 @@ const SignupPageFirst = ({navigation}: SignUpScreenFirstProp) => {
             </YStack>
           </XStack>
 
-          {/* Handle availability feedback */}
-          {isLoading && (
-            <Text color="green" fontSize={14}>
-              Checking...
-            </Text>
-          )}
-          {!isLoading && checkhandle?.isTaken && (
-            <Text color="red" fontSize={14}>
-              User handle is already in use.
-            </Text>
-          )}
-          {!isLoading && checkhandle?.isAvailable && (
-            <Text color="green" fontSize={14}>
-              User handle is available.
-            </Text>
-          )}
           {/* User Handle */}
           <XStack position="relative">
             <Input
@@ -419,6 +400,23 @@ const SignupPageFirst = ({navigation}: SignUpScreenFirstProp) => {
               <Icon name="person" size={20} color="#000" />
             </YStack>
           </XStack>
+
+          {/* Handle Availability Feedback */}
+          {isCheckingHandle && (
+            <Text color="$gray10" fontSize={14}>
+              Checking availability...
+            </Text>
+          )}
+          {!isCheckingHandle && handleAvailability && !handleAvailability.isAvailable && (
+            <Text color="red" fontSize={14}>
+              {handleAvailability.message}
+            </Text>
+          )}
+          {!isCheckingHandle && handleAvailability?.isAvailable && (
+            <Text color="green" fontSize={14}>
+              {handleAvailability.message}
+            </Text>
+          )}
 
           {/* Email */}
           <XStack position="relative">
